feat(named-spa-routes): compile any named location to href
namedLocationHref resolves against dummyRouteRecordsFromMetadata.
noteShowHref wraps it. routes.spec pins notebooks as a second compile
example besides note-show.
